Add tabs for feeds
Add tabs for feeds
Separate card list as a file from feed
Minor chages (main title, tab style)
